Add one flock + usleep protection GH #6
When sending in a short time multiple queries to the server...
it can happen that it reaches a state where he cannot recover...

This is a bug reported upstream...
in order to avoid to trigger this...
add a dummy flock mechanism to be more gentle with
the snapcast server.

The timeout added probably needs to move to configuration...
as this could depend on the network but better to hide it
for now.