Improve workaround for multiple queries - GH #6
This is only required in snapcast 0.10 as this
is fixed upstream by issue 200 in 0.11.

This commit is about providing a global protection
to any Telnet requests... loading the homepage.
As suggested by @airdrummingfool during code review.

This is slowing down all requests to avoid to trigger
the crash bug...

As it, it's still a pretty ugly workaround that
is not necessary with future snapcast server versions.

Long life to 0.11 and later.